Week-6/guardrails/Agent-Task/requirements.txt
boto3==1.42.32
chromadb==1.4.1
psycopg2-binary
fastmcp==2.14.3
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
langchain==1.2.6
langchain-aws==1.2.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-mcp-adapters==0.2.1
langchain-text-splitters==1.1.0
langgraph==1.0.6
langchain_postgres
langsmith==0.6.4
mcp==1.25.0
pypdf
langchain_tavily
langchain_ollama
python-docx
docx2txt
